+++ plugins/indentlist/plugin.js
@@ -221,11 +221,14 @@
 				}
 			}
 
+			if ( newList )
+				editor.fire( 'contentDomInvalidated' );
+
 			return true;
 		}
 
 		var selection = editor.getSelection(),
-			ranges = selection && selection.getRanges( 1 ),
+			ranges = selection && selection.getRanges(),
 			iterator = ranges.createIterator(),
 			range;